requirements.txt
asgiref
Django==4.2.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
DjangoUeditor==1.8.143
numpy==1.24.3
Pillow==9.5.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
